move modal button to stock folder
